+++ CHANGELOG.md
@@ -23,6 +23,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.3.1] - 2017-03-03
 ### Changed
+- AccountLevelIPDirectionalGroup: rename from AccountLevelIPDirectionalGroupDTO
+- AccountLevelGeoDirectionalGroup: rename from AccountLevelGeoDirectionalGroupDTO
+- ProbeAlertDataList: rename from ProbeAlertDataListDTO
+- ProbeAlertData: rename from ProbeAlertDataDTO
+- AccountList: rename from AccountListDTO
+- IPAddrRange: rename from IPAddrDTO
 - Client.NewRequest: shallow-copy BaseURL to avoid retaining modifications
 
 ## [1.3.0] - 2017-02-28